Adicion de Api Response y GlobalException handler

+++ foro/src/main/java/uni/fis/foro/dto/AsignarCategoriaForoDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class AsignarCategoriaForoDTO {
-    private Long idForo;
-    private Long idCategoria;
+    private Integer idForo;
+    private Integer idCategoria;
 }
